Refactor tool call middleware to streamline parameter transformation
- Updated the `transformParams` function in `tool-call-middleware.ts` to utilize a new `transformParams` utility, enhancing clarity and maintainability.
- Removed redundant code related to prompt processing and tool choice handling, simplifying the middleware's logic.
- Improved the overall structure of the middleware to better align with the new utility functions, ensuring more efficient tool call processing.

These changes enhance the readability and efficiency of the tool call middleware in the parser package.
